implements LFA feedback [Finishes 167730744]
